make some tasks
