+++ ctx_test.go
@@ -32,7 +32,6 @@ import (
 
 	"github.com/gofiber/fiber/v2/internal/bytebufferpool"
 	"github.com/gofiber/fiber/v2/internal/storage/memory"
-	"github.com/gofiber/fiber/v2/internal/template/html"
 	"github.com/gofiber/fiber/v2/utils"
 	"github.com/valyala/fasthttp"
 )
@@ -2735,127 +2734,6 @@ func Test_Ctx_Render(t *testing.T) {
 	utils.AssertEqual(t, false, err == nil)
 }
 
-// go test -run Test_Ctx_Render_Mount
-func Test_Ctx_Render_Mount(t *testing.T) {
-	t.Parallel()
-
-	sub := New(Config{
-		Views: html.New("./.github/testdata/template", ".gohtml"),
-	})
-
-	sub.Get("/:name", func(ctx *Ctx) error {
-		return ctx.Render("hello_world", Map{
-			"Name": ctx.Params("name"),
-		})
-	})
-
-	app := New()
-	app.Mount("/hello", sub)
-
-	resp, err := app.Test(httptest.NewRequest(MethodGet, "/hello/a", nil))
-	utils.AssertEqual(t, StatusOK, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-
-	body, err := ioutil.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "<h1>Hello a!</h1>", string(body))
-}
-
-// go test -run Test_Ctx_Render_Mount_ParentOrSubHasViews
-func Test_Ctx_Render_Mount_ParentOrSubHasViews(t *testing.T) {
-	t.Parallel()
-
-	engine := &testTemplateEngine{}
-	err := engine.Load()
-	utils.AssertEqual(t, nil, err)
-
-	engine2 := &testTemplateEngine2{}
-	err = engine2.Load()
-	utils.AssertEqual(t, nil, err)
-
-	sub := New(Config{
-		Views: html.New("./.github/testdata/template", ".gohtml"),
-	})
-
-	sub2 := New(Config{
-		Views: engine2,
-	})
-
-	app := New(Config{
-		Views: engine,
-	})
-
-	app.Get("/test", func(c *Ctx) error {
-		return c.Render("index.tmpl", Map{
-			"Title": "Hello, World!",
-		})
-	})
-
-	sub.Get("/world/:name", func(c *Ctx) error {
-		return c.Render("hello_world", Map{
-			"Name": c.Params("name"),
-		})
-	})
-
-	sub2.Get("/moment", func(c *Ctx) error {
-		return c.Render("bruh.tmpl", Map{})
-	})
-
-	sub.Mount("/bruh", sub2)
-	app.Mount("/hello", sub)
-
-	resp, err := app.Test(httptest.NewRequest(MethodGet, "/hello/world/a", nil))
-	utils.AssertEqual(t, StatusOK, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-
-	body, err := ioutil.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "<h1>Hello a!</h1>", string(body))
-
-	resp, err = app.Test(httptest.NewRequest(MethodGet, "/test", nil))
-	utils.AssertEqual(t, StatusOK, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-
-	body, err = ioutil.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "<h1>Hello, World!</h1>", string(body))
-
-	resp, err = app.Test(httptest.NewRequest(MethodGet, "/hello/bruh/moment", nil))
-	utils.AssertEqual(t, StatusOK, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-
-	body, err = ioutil.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "<h1>I'm Bruh</h1>", string(body))
-
-}
-
-func Test_Ctx_Render_MountGroup(t *testing.T) {
-	t.Parallel()
-
-	micro := New(Config{
-		Views: html.New("./.github/testdata/template", ".gohtml"),
-	})
-
-	micro.Get("/doe", func(c *Ctx) error {
-		return c.Render("hello_world", Map{
-			"Name": "doe",
-		})
-	})
-
-	app := New()
-	v1 := app.Group("/v1")
-	v1.Mount("/john", micro)
-
-	resp, err := app.Test(httptest.NewRequest(MethodGet, "/v1/john/doe", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-
-	body, err := ioutil.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "<h1>Hello doe!</h1>", string(body))
-}
-
 func Test_Ctx_RenderWithoutLocals(t *testing.T) {
 	t.Parallel()
 	app := New(Config{
@@ -3161,6 +3039,7 @@ func Test_Ctx_RestartRoutingWithChangedPathAndCatchAll(t *testing.T) {
 
 type testTemplateEngine struct {
 	templates *template.Template
+	path      string
 }
 
 func (t *testTemplateEngine) Render(w io.Writer, name string, bind interface{}, layout ...string) error {
@@ -3172,24 +3051,10 @@ func (t *testTemplateEngine) Render(w io.Writer, name string, bind interface{}, 
 }
 
 func (t *testTemplateEngine) Load() error {
-	t.templates = template.Must(template.ParseGlob("./.github/testdata/*.tmpl"))
-	return nil
-}
-
-type testTemplateEngine2 struct {
-	templates *template.Template
-}
-
-func (t *testTemplateEngine2) Render(w io.Writer, name string, bind interface{}, layout ...string) error {
-	if len(layout) == 0 {
-		return t.templates.ExecuteTemplate(w, name, bind)
+	if t.path == "" {
+		t.path = "testdata"
 	}
-	_ = t.templates.ExecuteTemplate(w, name, bind)
-	return t.templates.ExecuteTemplate(w, layout[0], bind)
-}
-
-func (t *testTemplateEngine2) Load() error {
-	t.templates = template.Must(template.ParseGlob("./.github/testdata2/*.tmpl"))
+	t.templates = template.Must(template.ParseGlob("./.github/" + t.path + "/*.tmpl"))
 	return nil
 }
