+++ group.go
@@ -18,33 +18,6 @@ type Group struct {
 	Prefix string
 }
 
-// Mount attaches another app instance as a sub-router along a routing path.
-// It's very useful to split up a large API as many independent routers and
-// compose them as a single service using Mount.
-func (grp *Group) Mount(prefix string, fiber *App) Router {
-	groupPath := getGroupPath(grp.Prefix, prefix)
-	groupPath = strings.TrimRight(groupPath, "/")
-	if groupPath == "" {
-		groupPath = "/"
-	}
-
-	// Support for configs of mounted-apps and sub-mounted-apps
-	for mountedPrefixes, subApp := range fiber.appList {
-		subApp.mountPath = groupPath + mountedPrefixes
-		grp.app.appList[groupPath+mountedPrefixes] = subApp
-	}
-
-	// Fill some fields of sub-app
-	fiber.parentApp = grp.app
-
-	// Execute onMount hooks
-	if err := fiber.hooks.executeOnMountHooks(grp.app); err != nil {
-		panic(err)
-	}
-
-	return grp
-}
-
 // Name Assign name to specific route.
 func (grp *Group) Name(name string) Router {
 	grp.app.mutex.Lock()
